+++ phpunit.xml.dist
@@ -20,7 +20,6 @@
     <testsuites>
         <testsuite name="Yii Definitions tests">
             <directory>./tests/Unit</directory>
-            <directory phpVersion="8.1" phpVersionOperator=">=">./tests/Php8_1</directory>
             <directory phpVersion="8.2" phpVersionOperator=">=">./tests/Php8_2</directory>
         </testsuite>
     </testsuites>
